refactor(tui): modal を開かないモードを keys/mode.py に分ける
keys/modal.py にコメント選択・削除確認・検索が同居しており、名前と中身が
合っていなかった。modal を開かず一覧の上でキーの解釈だけが変わるこれらを
mode.py に移し、modal.py は modal の開閉とその間のキーだけにする。

modal.py 322 行 -> 221 行 + mode.py 112 行。移動のみで挙動は変えていない。
